Updated the local state after following an account
refs https://linear.app/ghost/issue/AP-523

We want to preempt the Accept activity from our Follows, so we make the
assumption that it's succeeded. What this means is that we have to update our
`following`, `followingCount` as well as the fetched profile to set the
`isFollowing` property. This gives a more fluid user experience when following
accounts and keeps our state up to date.

Accounts where the follow request has to be accepted manually, are a little
trickier as we don't currently have easy access to the "requested but not
accepted state"