Refactored identity token generation into separate service
refs https://linear.app/ghost/issue/AP-500

The logic for generating identity tokens, whilst small, is something that we
don't want to duplicate - as it concerns security & access - so can easily
break interactions between different services. We're gonna need to use identity
tokens as part of the initialisation of the activitypub service, so this is
pulling it out preemptively for that use case

We shouldn't have logic inside of the endpoint controllers anyway, so this is
kinda general cleanup.