🐛 Fixed scheduled post datepicker marking prior day as selected
ref https://linear.app/ghost/issue/ONC-590

When choosing dates to schedule a post in the future, it could end up displaying the wrong selected date because it was accounting for local TZ. This doesn't make sense as we're displaying the site TZ in the picker itself, and that's the real TZ used for scheduling.

In short, this feels confusing and also is often incorrect/misleading, even though the scheduler in the background is correct. This should align those to make it more transparent.