🐛 Bulk adding tags created "duplicate" posts_tags rows from editor posts page
https://github.com/TryGhost/Ghost/issues/21414
The commit includes a migration which adds a unique index to posts_tags table and handles the related error when bulk adding posts.

The errors are ignored because the tags user wants to add are added/exist. This respects the editor's expectation as posts have the tags as the end result.