Add saveRaw method to LocalFilesStorage and LocalMediaStorage classes
ref https://linear.app/ghost/issue/CON-3/external-media-inliner-should-inline-more-file-types

The external media inliner requests each file as a buffer. Saving this works well for images, but not media or files as the `saveRaw` method is not available on those storage adapters.

This PR adds the same `saveRaw` method from `LocalImagesStorage` to two related classes, which enables saving buffers for all other supported file types