Added Sentry logging for email link clicks w/ bad member uuid (#21821)
no ref

I've discovered email link checkers appear to be using
falsified/scrambled uuids when testing links for a given site. It's
difficult to check against all sites, so instead we'll log to Sentry to
confirm this is the case.

I've put this behind config because I believe it will create a LOT of
entries, and may burden an already burdened workflow during peak
traffic.