Updated toggle for mobile navigation
Previously, the only way to close the navigation in Ghost Admin on mobile was to tap outside of the menu, on the background. With this change, you can also tap the 'More' button, or navigate to any other menu item, to have the menu overlaid on top close itself again.

fixes https://linear.app/ghost/issue/DES-993/the-slide-out-panel-inside-ghosts-admin-on-mobile-lacks-intuitive