🐛 Fixed mobile navigation for Admin (#21863)
Previously, the only way to close the navigation in Ghost Admin on
mobile was to tap outside of the menu, on the background behind it. With
this change, you can also tap the 'More' button or navigate to any other
menu item to have the menu close itself again.

fixes
https://linear.app/ghost/issue/DES-993/the-slide-out-panel-inside-ghosts-admin-on-mobile-lacks-intuitive

https://github.com/user-attachments/assets/db48a8cd-db4f-4118-9d6a-7b7b3d5e6236