🐛 Fixed post list not displaying correctly with long titles (#21892)
Previously, if you had posts with very long titles, the layout of the
post list would break, causing the analytics button and the metrics to
shift from row to row.

These changes address that, and make the list render cohesively
regardless of the length of a post's title, across resolutions.

fixes
https://linear.app/ghost/issue/DES-610/post-list-breaks-with-really-long-titles