Handle null values when stripping HTML from post content in admin-x-activitypub
refs [TryGhost/ActivityPub#245](https://github.com/TryGhost/ActivityPub/pull/245)

In [TryGhost/ActivityPub#245](https://github.com/TryGhost/ActivityPub/pull/245)
we changed the service to allow for `null` values for the `content` field. This
means we could potentially be passing `null` values to `stripHtml` which would
cause the app to crash. This commit ensures we always pass a string to `stripHtml