Removed autosaving of tags from gh-post-settings-menu
ref https://linear.app/ghost/issue/ONC-756
ref d91869e6406b7b1f0790cb34c92d2b8c992f3465

Rather than forcing a save on every tag change, we want to defer to the
autosave functionality which will debounce saves, reducing the number of API
calls when modifying tags on a post. This does mean that changing tags for
existing posts will require an explicit save, as the autosave functionality
does not run for published posts.