Fixed leftover state in invites browser tests (#22136)
no issue

- The invite tests were signing out as the owner user and then signing
in as a user with different privileges. At the end of the test, it was
left in this state, so the next file run by that playwright worker was
still logged in as a user with lower than owner privileges so e.g. the
settings couldn't be accessed.
- This change signs out from the newly created user and signs back in as
the owner user after each of the invites tests.