ActivityPub Shade integration (#22150)
ref https://linear.app/ghost/issue/AP-700/explore-ui-development-in-ap

This PR is about integrating ActivityPub with the new design system, so that we can start using Shade in a real world application. In this early version the TailwindCSS config is taken from Shade but the custom styles of Shade and the AdminX Design System are merged. This is needed so that we can keep using the old components without breaking functionality.