Fixed e2e test
Note: we previously had no test coverage with the `checkout.session.completed` handler in the case of an existing member. This had to modify a couple pieces to allow for that.

Because we're editing the subscriptionCreatedEvent, the memberCreatedEvent info needs to be skipped as it may not match (which is expected).